Implemented documents.

+++ treeflaskapp/blueprint_documents.py
@@ -1,7 +1,7 @@
-from flask import Blueprint, request, render_template, redirect, url_for, g, flash
+from flask import Blueprint, request, render_template, redirect, url_for, g, flash, jsonify, current_app
 from flask_login import login_required, current_user
-from .models import Document, Person, Place, Event, db
-from .forms import DocumentForm
+from .models import Document, File, Person, Place, Event, db
+from .forms import DocumentForm, FileForm
 from werkzeug.utils import secure_filename
 import os
 from datetime import datetime
@@ -14,10 +14,9 @@ def handle_file_upload(file, id, is_image=False):
         if is_image:
             ALLOWED_EXTENSIONS = {'jpg', 'png', 'jpeg', 'gif', 'JPG', 'JPEG'}
         else:
-            ALLOWED_EXTENSIONS = {'jpg', 'png', 'jpeg', 'gif', 'JPG', 'JPEG', 'txt', 'pdf', 'doc', 'docx', 'xls', 'xlsx',
-                                  'ppt', 'pptx', 'csv', 'rtf', 'odt', 'ods', 'odp', 'zip', 'rar', 'tar', 'gz',
-                                  '7z'}  # добавить другие расширения файлов по мере необходимости
-        return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+            ALLOWED_EXTENSIONS = None  # allow all extensions
+        return '.' in filename and (
+                    ALLOWED_EXTENSIONS is None or filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS)
 
     if file and allowed_file(file.filename):
         filename = secure_filename(file.filename)
@@ -38,11 +37,13 @@ def handle_file_upload(file, id, is_image=False):
 
 
 @documents.route('/user/<username>/documents')
+@login_required
 def documents_view(username):
     if username == current_user.username:
         documents = Document.query.filter_by(user_id=current_user.id).all()
-        form = DocumentForm(user_language=g.user_language)  # create an instance of your form
-        return render_template('documents.html', documents=documents, form=form)
+        form = DocumentForm(user_language=g.user_language)  # create an instance of your DocumentForm
+        file_form = FileForm(user_language=g.user_language)  # create an instance of your FileForm
+        return render_template('documents.html', documents=documents, form=form, file_form=file_form)
     else:
         return "Unauthorized", 403
 
@@ -50,24 +51,29 @@ def documents_view(username):
 @login_required
 def create_document(username):
     form = DocumentForm(user_language=g.user_language)
+    file_form = FileForm(user_language=g.user_language)
     if form.validate_on_submit():
         try:
             icon_path = None
             if 'icon' in request.files and request.files['icon'].filename != '':
                 icon_path = handle_file_upload(request.files['icon'], is_image=True, id=current_user.id)
 
-            file_path = None
-            if 'file_path' in request.files and request.files['file_path'].filename != '':
-                file_path = handle_file_upload(request.files['file_path'], is_image=False, id=current_user.id)
-
             document = Document(user_id=current_user.id,
                                 name=form.name.data,
                                 description=form.description.data if form.description.data else None,
                                 date=form.date.data if form.date.data else None,
-                                file_path=file_path,
                                 comment=form.comment.data if form.comment.data else None,
                                 icon=icon_path if icon_path else None)
             db.session.add(document)
+            db.session.commit()
+
+            if 'file_path' in request.files:
+                for file in request.files.getlist('file_path'):
+                    if file.filename != '':
+                        file_path = handle_file_upload(file, is_image=False, id=current_user.id)
+                        file = File(file_path=file_path, document_id=document.id)
+                        db.session.add(file)
+
             db.session.commit()
             flash_message = 'Document created successfully!' if g.user_language == 'en' else 'Документ успешно создан!'
             flash(flash_message, 'success')
@@ -77,19 +83,19 @@ def create_document(username):
             flash_message = 'An error occurred while creating the document: {}'.format(e) if g.user_language == 'en' \
                 else 'Произошла ошибка при создании документа: {}'.format(e)
             flash(flash_message, 'error')
-    return render_template('create_document.html', form=form, username=username)
+    return render_template('create_document.html', form=form, file_form=file_form, username=username)
 
 @documents.route('/user/<username>/edit_document/<int:id>', methods=['GET', 'POST'])
 @login_required
 def edit_document(username, id):
     document = Document.query.get(id)
     if document is None or document.user_id != current_user.id:
-        return "Unauthorized", 403
+        abort(403)
 
     form = DocumentForm(obj=document, user_language=g.user_language)
+    file_form = FileForm(user_language=g.user_language)
     if form.validate_on_submit():
         try:
-            # icon_path = None
             if 'icon' in request.files and request.files['icon'].filename != '':
                 icon_path = handle_file_upload(request.files['icon'], is_image=True, id=current_user.id)
                 document.icon = icon_path  # update image_file field with full path
@@ -97,9 +103,14 @@ def edit_document(username, id):
             document.name = form.name.data if form.name.data else None
             document.description = form.description.data if form.description.data else None
             document.date = form.date.data if form.date.data else None
-            document.file_path = form.file_path.data if form.file_path.data else None
             document.comment = form.comment.data if form.comment.data else None
-            # document.icon = form.icon.data if form.icon.data else None
+
+            if 'file_path' in request.files:
+                for file in request.files.getlist('file_path'):
+                    if file.filename != '':
+                        file_path = handle_file_upload(file, is_image=False, id=current_user.id)
+                        file = File(file_path=file_path, document_id=document.id)
+                        db.session.add(file)
 
             db.session.commit()
             flash_message = 'Document updated successfully!' if g.user_language == 'en' else 'Документ успешно обновлен!'
@@ -109,7 +120,8 @@ def edit_document(username, id):
             db.session.rollback()
             flash('An error occurred while updating the document: {}'.format(e), 'error')
 
-    return render_template('edit_document.html', form=form, username=username, document=document)
+    return render_template('edit_document.html', form=form, file_form=file_form, username=username, document=document)
+
 
 @documents.route('/user/<username>/delete_document/<int:id>', methods=['POST'])
 @login_required
@@ -119,6 +131,27 @@ def delete_document(username, id):
         return "Unauthorized", 403
 
     try:
+        if document.icon:
+            # If the document's icon is the same as the filename, set it to None
+            file_path = os.path.join(current_app.root_path, 'static', document.icon);
+
+            if os.path.exists(file_path):
+                os.remove(file_path)
+
+            document.icon = None
+            db.session.commit()
+
+        # Delete the files associated with the document
+        for file in document.files:
+            # Construct the full file path
+            file_path = os.path.join(current_app.root_path, 'static', file.file_path)
+            # Check if the file exists and delete it
+            if os.path.exists(file_path):
+                os.remove(file_path)
+            # Delete the file record from the database
+            db.session.delete(file)
+
+        # Delete the document
         db.session.delete(document)
         db.session.commit()
         flash_message = 'Document deleted successfully!' if g.user_language == 'en' else 'Документ успешно удален!'
@@ -135,26 +168,40 @@ def delete_document(username, id):
 @login_required
 def delete_file(username, id):
     data = request.get_json()
-    filename = data.get('filename')
+    filename = data.get('filename'); print(f"filename: {filename}")
+    file_path = data.get('file_path'); print(f"file_path: {file_path}")
 
-    print('Received request:', request)  # Print the entire request
-    print('Received data:', data)  # Print the received JSON data
-    print('Received filename:', filename)  # Print the received filename
-
-    if filename:
+    if filename or file_path:
         document = Document.query.get(id)
-        if document and document.icon == filename:
-            # If the person's image_file is the same as the filename, set it to None
-            document.icon = None
-            db.session.commit()
+        if document:
+            if document.icon == filename and filename is not None and filename != "":
+                # If the document's icon is the same as the filename, set it to None
+                file_path = os.path.join(current_app.root_path, 'static', filename);
 
-            file_path = os.path.join(documents.root_path, 'static', filename)
+                if os.path.exists(file_path):
+                    os.remove(file_path)
 
-            if os.path.exists(file_path):
-                os.remove(file_path)
+                document.icon = None
+                db.session.commit()
 
-            return jsonify({'message': 'File deleted.'}), 200
+            file_to_delete = None
+            for file in document.files:
+                if file.file_path == file_path:
+                    file_to_delete = file
+                    break
+
+            if file_to_delete:
+                # If the document's file path is the same as the filename, set it to None
+                file_path = os.path.join(current_app.root_path, 'static', file_path); print(f"file_path: {file_path}")
+
+                if os.path.exists(file_path):
+                    os.remove(file_path)
+
+                db.session.delete(file_to_delete)
+                db.session.commit()
+
+                return jsonify({'message': 'File deleted.', 'success': True}), 200
         else:
-            return jsonify({'message': 'File not found.'}), 404
+            return jsonify({'message': 'Document not found.'}), 404
     else:
         return jsonify({'message': 'No filename provided.'}), 400
